diffinsights_web: Fix "Period:" selector not updating
Since 03d4d8a1 the "Period:" from date selector is counted from the last
date in annotation data (last date in repo/dataset), instead of being
counted from the current date (i.e. today).

This means that after chaning the dataset, it should get update.  This
commit ensures that it is done, by resetting it if necessary - it is
needed to not have date that is not on the list of slectable options,
and not have the date which is outside date range for the current
repository / dataset.

NOTE that this means that the 'from' date gets reset when repository is
changed, as opposed to how the app behaved before that change, where it
remained the same, and you can switch between repositores.  Perhaps the
past behavior could be offered as an option.