diffinsights_web: Introduce datastore/linesstats.py
Currently only the data reading part is implemented.  As this datastore
is intended to be an auxiliary data store, to augment timeline data
store with data required to create Sankey diagram, it is intended to try
to find a companion data file to the one used by the timeline data
store.

Also added some basic tests, which are skipped if .[web] prerequisites
are not available.