diffinsights_web: Add LineStatsDataStore.sorted_changed_files()
Probably it would be changed to be a function, in order to make it
easier to make reactive expression out of it.  It is not something
specific to the LineStatsDataStore class...

This function will be used to implement cutoff, to limit Sankey diagram
to top N most changed files (istead of using all files).