diffinsights_web: Add reduce_sankey_thin_out()
Another function to reduce number of nodes in Sankey diagram by
assigning their contributions to the parent node (there is only one
parent node).

It is quite conservative, and not the best at removing nodes.
Need rethinging, so that it would be able to compress / reduce
chain of nodes (with all edges on the chain being thin edges).