diffinsights_web: Add src/diffinsights_web/apps/author.py
The code is copy from notebooks/panel/01-timeline.ipynb, adjusted to run
as a Python script and not in a Jupyter notebook.  It tries to re-use
what it can from other modules in diffinsights_web.

No care was taken to extract common code, or to split the code into
smaller modules; the code uses functional approach that utilizes
reactive expressions, with some @pn.cache sprinkled in.  It does not use
class based approach from "Structure with a DataStore"
https://panel.holoviz.org/tutorials/intermediate/structure_data_store.html
that 'diffinsights_web/apps/contributors.py' uses.