diffinsights_web: Fix selection of author in apps/author.py
For the 'tensorflow.timeline.purpose-to-type.json' data file it was
possible to use 'bug_id' as a proxy for the author, because this file
was generated with two runs of `diff-annotate from-repo ...`, each
run selecting contributions from different author.

Use instead 'author.name' column to select the author ('author.email'
would also be a good choice).

TODO: make it configurable.