diffinsights_web: Don't use rescale_n_mod_widget
Lines are counted as modified if there are added and removed lines in
a single change group.  In unified diff format (that Git uses), each
changed line is represented as deletion of old version of the line,
and adding new version of the line.  Thus one removed line corresponds
to one removed line and one added line, that is two changed lines.

Therefore, on -/+ plot, with added lines going up, and removed lines
going down, itis better to show number of modified lines on both:
they denote number of added lines that are counted as modified lines,
and similarly for removed lines.

Also, always use "mod" as label.