diffinsights_web: Introduce and use plot_periodicity_heatmap()
This plot is intended to show perodicity in users behavior.  Currently
it shows heatmap of number of commits, grouped by hour and by day of
week, with histograms for number of commits made in a give hour, etc.

The quality of results depends on whether the information about author
timezone is correct or not.