Add .python_version (for Heroku)
Heroku by default uses the same Python version as the last build.
It was Python 3.12.7, for which there is a security update:

 !     Warning: A Python security update is available!
 !
 !     Upgrade as soon as possible to: Python 3.12.8
 !     See: https://devcenter.heroku.com/articles/python-runtimes

It is recommended to specify only the major version (e.g. 3.13),
so that your app is automatically built with the latest Python security
patches.

For simplicty, stay at the same major version, even though 3.13
is available.