diffinsights_web: Use DataFrame as source for Sankey diagram
This will allow for easier manipulation of labels and styles; the
alternative would be to use HoloViews kdims and vdims, with tuples
as source.

Currently colors of nodes and colors of edges are quite arbitrary,
just so that you can distinguish between them.

Specified tools to use, to get rid of annoing wheel-zoom tools
(which makes trying to use wheel to scroll the page very annoying).

.
NOTE: it turns out that replacing directory names with only their last
component does not work for hv.Sankey, where that name is the identifier
of node... and in Java project there are subdirectories of different
directories with the same name, e.g.:
  - src/tests/java/...
  - src/main/java/...

In next commit we will go back to use the full name.