diffinsights_web: Move Sankey widgets to sidebar, prepare for disabling
In the next commit, the WidgetBox with all widgets controlling the
creation and behavior of Sankey diagram will be disabled as whole when
other plot is shown (not Sankey diagram).

Alternative would be to come with some wahy to put Sankey diagram
widgets in a row above the plot itself.