docs: Use "Automatic code reference pages" mkdocstrings recipe
`mkdocstrings` allows to inject documentation for any object into Markdown
pages.  Then, the `mkdocs-gen-files` plugin exposes utilities to generate
files at build time, here we use it to generate code reference pages.
Together with gen-files plugin, we use `mkdocs-literate-nav`, which allows
to specify the whole navigation, or parts of it, into Markdown pages,
as plain Markdown lists - which can be generated by a script.  Finally,
`mkdocs-section-index` allows to render `__init__` modules documentation
inside the section itself.

One change we did was to comment out set_edit_path line, because this
feature is turned off in mkdocs.yml.

.
It turns out, however, that the example scripts in the recipe doesn't
handle correctly the case where there are two main modules in src/.
In PatchScope we have two: diffannotator, and diffinsights_web.
The documentation for second (where, incidentally, we have fewer
docstrings) overwrites, at least partially, the documentation for the
first.

This would have to be fixed.