docs: Add "Flow from path to line type (Sankey diagram)" to contributors_graph.md
This documents current state of Sankey diagram (Sankey plot) support in
diffinsights_web, as of version 0.4.1.  It is not complete, and not
finished.  Among other things, there is no support for selecting period
of time ("From"), nor for selecting author (for per-author plots on authors
grid).

Currently the Sankey diagram is plotted using HoloViews (low-level
plotting library) with the interative Bokeh backend.  There are however
other options, with their advantages and disadvantages:
- Plotly (interactive)
- Mermaid.js (with gradient, non-interactive), requires additional
  support: awesome-panel/panel-mermaid or oruelle/md_mermaid
- pySankey (mentioned in python-graph-gallery.com)
- from PyPI: Irene-Sankey, mpl_sankey, Sugikey (interactive via Bokeh),
  SankeyFlow, pSankey, pySankey2, Sankeyfy (for Plotly API, maybe as
  inpiration), floWeaver + ipysankeywidget