experiments/01-compare_annotations.ipynb: per-line consensus per user/task
Compute percentage of cases which have consensus and agree with consensus
on a per annotation task (per user, i.e. manual annitator, per changed line).

This part of the "Per line-type consensus (addition)" subsection shows that
the task suggested by the reviewer/review #110C of the "HaPy-Bug – Human
Annotated Python Bug Resolution Dataset" paper is not well defined.

> The authors report interesting results, indicating that only a small
> proportion of lines (≤ 6%) required manual intervention after automatic
> annotations. More details on the nature of these modifications would
> strengthen the analysis. Specifically, it would be helpful to know
> if the automatic annotations were primarily inaccurate in specific
> categories (e.g., documentation) or whether they lacked context for
> certain types of code changes. Since the automation relies on simple
> syntax rules, it might miss complex bug patterns, which requires
> further discussion.

TODO: On a per-line basis, check if automated annotation agrees with
consensus, and if there is consensus for the line - grouped by automated
annotation label for the changed line.