benchmarks: Benchmark main components of the 'contributors' app
This adds the test_contributors_steps_performance() test, in which
during its creation main steps (main components) of running the
'contributors' app were benchmarked, and benchmark results were saved
as comments.

This shows which parts are worth caching, and which do not need it.

While at it, fix the ordering of imports.  All diffinsights_web imports
must be after import of 'panel' and 'param' via pytest.importorskip()!