benchmarks: Make get_timeline_df() use first repo on KeyError
This is workaround for current inability to create a test benchmark
with data different from the default 'hellogitworld' example data.

Trying to change the input JSON data file leads to error: JSON file
is changed, but unexpectedly repository list and repo parameter
is not adjusted - contrary to what happens for the app when run.
See one of the previous commit in "benchmarks" series.

This commit therefore makes get_timeline_df() simply use first
entry in the `timeline_data` dict if `repo` is not present among
its keys.

This change is made as a separate commit to make it easy to reverse.