benchmarks: Test which pandas save format has best performance
We are interested mainly in the deserialiation (i.e. read) performance.
This is planned to be used to cache computed dataframe(s) on disk.

Results are presented in the comment before first benchmarking test.
They agree with experiments peformed by Ilia Zaitsev in 2019:
https://medium.com/towards-data-science/the-best-format-to-save-pandas-data-414dca023e0d
(for Towards Data Science blog) that feather format has almost best
preformance, while being safer than pickle.

Note that, as opposed to Parquet, the Arrow / Feather format was not
originally intended for long-term storage.