benchmarks: Add 'pyarrow' to [web] dependencies, for caching
The 'pyarrow' library needs to be installed for Pandas to be able to run
the pd.read_feather(<file>) function to read cached DataFrame(s).

TODO?: use caching only if 'pyarrow' is installed, but do not use it as
a hard requirement (i.e. it can be installed, but is not required for
diffinsights_web to work).