benchmarks: Temporarily remove perspective panes from contributors app
This is intended to check how much the performance of the contributors
web app (part of diffinsights_web) is affected by pn.Tabs with JSON
viewer and various perspective panes.

Putting this change in a separate commit makes it so it can be
easily reverted.  TODO: move commented out functionality to separate
source file, and make it into a separate web app, putting link to
it at the bottom of the contributors page (perhaps as an option,
offer it at the bottom of contributors page, like it was before).