benchmarks: Add results for no perspective panes in 'contributors' app
Add to tests/test_app_contributors_performance.py (as comments)
results of benchmarks of the contributors app for 'hellogitworld',
'qtile' and 'tensorflow' repos after removing perspective panes
in the previous commit.

Those are trigger-based benchmark results for all repos, and
manual stopwatch results for 'qtile' repo - which got strange...
with each switch taking longer and longer for the spinner to stop
spinning.