benchmarks: Temporarily remove sidebar from contributors app
This is intended to check how much the performance of the contributors
web app (part of diffinsights_web) is affected by rendering widgets
in the sidebar and sidebar itself.

Putting this change in a separate commit makes it so it can be
easily reverted (which it probably will be done).

TODO: check if _hiding_ the sidebar affects performance.