diffinsights_web: Move data exploration to 'explore_repo' app
This is straight movement of JSON viewer and Perspective panes from
'contributors' app to newly created 'explore_repo' app, and copy
enough code to be able to run it.

TODO: Adjust and simplify copied code in explore_repo.py so that
it does not bring things that are not used for exploration.

TODO: Display JSON data, even if there is cached data, and
'contributors' app does not need to load it.