diffinsights_web: Extract path_to_name() function
This function is used to turn full path to input JSON file to the key
that is used to select said file.  Currently this is the part of
basename up to the first '.' - a convention introduced in an earlier
commit 377b8e18.

With this it would be easy to use this key as query parameter.