diffinsights_web: Use author's e-mail to select author in 'author' app
Not in all cases the name of the author (contributor) is given a name,
sometimes what is given is pseudonym; it doesn't look nice if we have
mixture of proper names and pseudonyms.

Unfortunately in query params are passed URI-encoded, for example
'tycho@tycho.ws' is present in URL as 'tycho%40tycho.ws'.