diffinsights_web: Use 'contributors' as index in Dockerfile
This follows the setup for multi-page apps from the documentation:
https://panel.holoviz.org/how_to/streamlit_migration/multipage_apps.html

The consequence is that with this configuration both of the following
URLs points to the same page:
  http://localhost:7860/
  http://localhost:7860/contributors

Solving this would probably involve using other server than Bokeh, for
example using FastAPI, or something like that.

We need to run both 'contributors' and 'author' apps to be able to
provide working hyperlinks from 'contributors' to 'author' pages
for individual authors.