diffinsights_web: Update repos widget when 'repo' is read from URL param
In the case where there is 'repo' query parameter set, which value
is the key that defines input JSON file, the widget that denotes part
of data from that file that we want to analyze also needs to be updated.

Without this change, when reloading page where 'repo' parameter is set,
the second widget in the sidebar gets empty value.