panel/03-author_Sankey.ipynb: Add "Mermaid.js plots" to diagram simplification
Mermaid support for Sankey flow diagrams is in beta, but it is the only
tool where turning on gradient color for edges is easy (which is perhaps
possible in Plotly / Plotly Express, but might not be possible for
HoloViews).

Try to create a nice, readable example that one can put as an example
figure in paper.