README.md: A few further improvements to the flow
Add link to the demo of the web application at the top of README.md,
where it can be easily found.

Enhance a bit description of how to run the web application without
cloning the repository.  Give link to open to see the just started
web application.

Other minor improvements.