diffinsights_web: Do not use TimelineDataStore for MermaidSankeyPlot
Because the 'author' app, that is to use MermaidSankeyPlot, do not
use TimelineDataStore, change MermaidSankeyPlot to not use it either,
but to use the same parameters that 'author' app uses to extract data.

Without this change the simple check failed - the widget did not
show correct number of changed lines.