diffinsights_web: Toggle between full and short Sankey config
In the 'author' app, add a toggle to the "Sankey flow diagram" widget
box that allows to use full configuration that Mermaid.js supports
for Sankey plots, not only selected subset of them.

The solution chosen is not the best, but it works.  This change is
intended to make it easier to create beautiful and readable Sankey
flow diagrams, e.g. for the paper.