diffinsights_web: Add some prints and watchers for debugging issues
It turns out that for some reason find_repos(), which is downstream
in the pipeline from the select_file_widget, is called 3 times (!)
instead of once for each change to value of select_file_widget.

@pn.cache has problems, because parameter to find_repos() is the
whole dict(), which would be used as key to cache - perhaps after
serialization - which is slow.