diffinsights_web: Remove "workaround" for repos widget empty on reload
This "workaround" is no longer needed, as the issue it tried to solve
was resolved by the previous commit.  Additionally, it set 'options' and
'value' fields for repos widget to something that was _not_ reactive
expression, thus taking repos widget out of widget pipeline.

So this commit fixes the issue of repos widget not updating after
value of files widget changed.