Add *.service unit file for systemd, runs web app dockerized
This systemd unit file assumes that ncusi/patchscope:latest image
is present locally on the system.

As of now, that image was not yet submitted to a container registry
like DockerHub (docker.io); when it will get submitted, you can
always download the image on restrting the web dashboard service - after
uncommenting some lines in docker.patchscope-web.service file.

Untested!