diffinsights_web: Default to 'qtile' repository, if available
When not provided with the repository / input JSON data file with the
'repo' query parameter in URL, use 'qtile' repository as the default,
if that repository is available.

Before this commit the default was the repository that happened to be
first in the DATASET_DIR directory (usually 'hellogitworld').

The 'qtile' repository was selected because it is nice default.