experiments/01-compare_annotations.ipynb: Remove SmartSHARK / Herbold et al.
Remove the whole "Comparison with line annotations in the Herbold et al.
dataset (hunk_labels.json only)" section, as the comparison there was
based on a faoulty assumption on how changes are represented in the
hunk_labels.json data file, and how that representation corresponds
to the diff of changes.

The correct comparison is present in a separate notebook, namely
experiments/02-compare_annotations_Herbold.ipynb

The state before the change was tagged with 'notebooks-before_experiments_cleanup'.