docs: Switch from mkdocs-typer to mkdocs-typer2
The mkdocs-typer plugin hasn't been updated in over a year, and there
have been a number of changes to Typer since then. One important change
is that Typer now has its own documentation generation system via the
`typer <module> utils docs` command. The mkdocs-typer2 plugin simply
leverages that system to generate the documentation for your Typer CLIs.

There were some problems with mkdocs-typer plugin because of the way
the 'typer' library is used in this project, and the workaround worked
only partially:
  https://github.com/bruce-szalwinski/mkdocs-typer/issues/18

Note that mkdocs-typer2 is plugin, while mkdocs-typer was markdown
extension.