docs: Rewrite links in included Markdown files with custom hook
Using README.md files to create basis of the MkDocs-based project
documentation with the help of mkdocs-include-markdown-plugin has one
disadvantage: links to other READMEs and other files from those
README.md files work for them when browsed on GitHub, but lead to 404
within the generated documentation.  The 'rewrite-relative-urls' option
helps, but is not a panacea.

For example, notebooks/README.md file contains link to main README.md.
This gets turned into ../README.md when include-markdown in notebooks.md
gets processed.  This needs to be adjusted to point to index.md (which
includes README.md).

Those README.md files also contain links to directories and non-Markdown
files.  Those needs to be turned in full GitHub URLs to directory (tree)
or file (blob) in question.

.
There is 'mkdocs-redirects' plugin to create page redirects, but it is
about actually creating a new page where an old page was that redirects
to the new page using JavaScript redirect.  It is intended for moved
and renamed pages, rather than for fixing links.
  https://github.com/mkdocs/mkdocs-redirects

Therefore a MkDocs hook was created, hooking `on_page_markdown` event,
and doing search and replace on links (using regular expressions).
The script with the hook is currently at scripts/my_hooks.py.
  https://www.mkdocs.org/user-guide/configuration/#hooks
  https://www.mkdocs.org/dev-guide/plugins/#on_page_markdown

While at it, the hook also checks for non HTTPS external links.