utils.git: Add close_batch_command() method to GitRepo class
This method, and the use of the finalizer object with weakref.finalize()

  https://docs.python.org/3/library/weakref.html#weakref.finalize
  https://docs.python.org/3/library/weakref.html#finalizer-objects

is here to be able to close the persistent subprocess to `git cat-file`,
and to avoid

  OSError: [Errno 24] Too many open files

exception when creating many GitRepo objects, and using methods that
access the `batch_command` cached property in a Python script.

Tried to test this new method and its internal implementation
in tests/test_utils_git.py

.
In the actual case this was the problem (above mentioned OSError
exception) happened in the following line:

    # emits SHA-1 identifier if object is found in the repo; otherwise, errors out
    process = subprocess.run(cmd, capture_output=True, check=True)

This line comes from `to_oid(obj)` method in the GitRepo class.