utils.git: Fix get_patched_file_mode() and test the fix
The problem was handling file mode detection for files in pre-image
(i.e., source) for file deletion: the extended diff header is

  deleted file mode <octal mode>

and not

  old file mode <octal mode>

For post-image and added files it worked correctly, because for
adding a new file the extended diff header is

  new file mode <octal mode>

which uses the same prefix as extended diff header for mode change

  old mode <source octal mode>
  new mode <target octal mode>

.
Add a new test in tests/test_utils_git.py that checks that
`get_patched_file_mode(<changed file>, side=DiffSide.PRE)`
works when <changed file> is being deleted.

While at it, add new data file and rename test data file to match.